refactor(navigation): share server-safe primitives
