+++ packages/vinext/src/server/app-browser-state.ts
@@ -48,7 +48,6 @@ export {
   createInitialBfcacheIdMap,
   createNextBfcacheIdMap,
   preserveBfcacheIdsForMergedElements,
-  type BfcacheStateKeyMap,
 } from "./app-bfcache-identity.js";
 
 export {
