+++ packages/vinext/src/shims/navigation.ts
@@ -63,6 +63,52 @@ import {
   getPendingAppRouterScrollIntent,
   type AppRouterScrollIntent,
 } from "./app-router-scroll-state.js";
+import {
+  clearClientHydrationContext,
+  getBfcacheIdMapContext,
+  getBfcacheSegmentIdContext,
+  getLayoutSegmentContext,
+  getNavigationContext,
+  registerServerInsertedHTMLCallback,
+  type NavigationContext,
+} from "./navigation-context-state.js";
+
+export {
+  type NavigationContext,
+  type NavigationStateAccessors,
+  type SegmentMap,
+  GLOBAL_ACCESSORS_KEY,
+  ServerInsertedHTMLContext,
+  _registerStateAccessors,
+  clearServerInsertedHTML,
+  flushServerInsertedHTML,
+  getBfcacheIdMapContext,
+  getBfcacheSegmentIdContext,
+  getLayoutSegmentContext,
+  getNavigationContext,
+  renderServerInsertedHTML,
+  setNavigationContext,
+} from "./navigation-context-state.js";
+
+export {
+  BailoutToCSRError,
+  DynamicServerError,
+  HTTP_ERROR_FALLBACK_ERROR_CODE,
+  RedirectType,
+  decodeRedirectError,
+  forbidden,
+  getAccessFallbackHTTPStatus,
+  isBailoutToCSRError,
+  isDynamicServerError,
+  isHTTPAccessFallbackError,
+  isNextRouterError,
+  isRedirectError,
+  notFound,
+  permanentRedirect,
+  redirect,
+  unauthorized,
+  unstable_rethrow,
+} from "./navigation-errors.js";
 
 // ─── Layout segment context ───────────────────────────────────────────────────
 // Stores the child segments below the current layout. Each layout wraps its
@@ -70,105 +116,8 @@ import {
 // (including route groups, with dynamic params resolved to actual values).
 // Created lazily because `React.createContext` is NOT available in the
 // react-server condition of React. In the RSC environment, this remains null.
-// The shared context lives behind a global singleton so provider/hook pairs
-// still line up if Vite loads this shim through multiple resolved module IDs.
-const _LAYOUT_SEGMENT_CTX_KEY = Symbol.for("vinext.layoutSegmentContext");
-const _SERVER_INSERTED_HTML_CTX_KEY = Symbol.for("vinext.serverInsertedHTMLContext");
-const _BFCACHE_ID_MAP_CTX_KEY = Symbol.for("vinext.bfcacheIdMapContext");
-const _BFCACHE_SEGMENT_ID_CTX_KEY = Symbol.for("vinext.bfcacheSegmentIdContext");
-
-/**
- * Map of parallel route key → child segments below the current layout.
- * The "children" key is always present (the default parallel route).
- * Named parallel routes add their own keys (e.g., "team", "analytics").
- *
- * Arrays are mutable (`string[]`) to match Next.js's public API return type
- * without requiring `as` casts. The map itself is Readonly — no key addition.
- */
-export type SegmentMap = Readonly<Record<string, string[]>> & { readonly children: string[] };
-
-type _LayoutSegmentGlobal = typeof globalThis & {
-  [_LAYOUT_SEGMENT_CTX_KEY]?: React.Context<SegmentMap> | null;
-  [_SERVER_INSERTED_HTML_CTX_KEY]?: React.Context<
-    ((callback: () => unknown) => void) | null
-  > | null;
-  [_BFCACHE_ID_MAP_CTX_KEY]?: React.Context<Readonly<Record<string, string>> | null> | null;
-  [_BFCACHE_SEGMENT_ID_CTX_KEY]?: React.Context<string | null> | null;
-};
-
-// ─── ServerInsertedHTML context ────────────────────────────────────────────────
-// Used by CSS-in-JS libraries (Apollo Client, styled-components, emotion) to
-// register HTML injection callbacks during SSR via useContext().
-// The SSR entry wraps the rendered tree with a Provider whose value is a
-// callback registration function (useServerInsertedHTML).
-//
-// In Next.js, ServerInsertedHTMLContext holds a function:
-//   (callback: () => React.ReactNode) => void
-// Libraries call useContext(ServerInsertedHTMLContext) to get this function,
-// then call it to register callbacks that inject HTML during SSR.
-//
-// Created eagerly at module load time. In the RSC environment (react-server
-// condition), createContext isn't available so this will be null.
-
-function getServerInsertedHTMLContext(): React.Context<
-  ((callback: () => unknown) => void) | null
-> | null {
-  if (typeof React.createContext !== "function") return null;
-
-  const globalState = globalThis as _LayoutSegmentGlobal;
-  if (!globalState[_SERVER_INSERTED_HTML_CTX_KEY]) {
-    globalState[_SERVER_INSERTED_HTML_CTX_KEY] = React.createContext<
-      ((callback: () => unknown) => void) | null
-    >(null);
-  }
-
-  return globalState[_SERVER_INSERTED_HTML_CTX_KEY] ?? null;
-}
-
-export const ServerInsertedHTMLContext: React.Context<
-  ((callback: () => unknown) => void) | null
-> | null = getServerInsertedHTMLContext();
-
-/**
- * Get or create the layout segment context.
- * Returns null in the RSC environment (createContext unavailable).
- */
-export function getLayoutSegmentContext(): React.Context<SegmentMap> | null {
-  if (typeof React.createContext !== "function") return null;
-
-  const globalState = globalThis as _LayoutSegmentGlobal;
-  if (!globalState[_LAYOUT_SEGMENT_CTX_KEY]) {
-    globalState[_LAYOUT_SEGMENT_CTX_KEY] = React.createContext<SegmentMap>({ children: [] });
-  }
-
-  return globalState[_LAYOUT_SEGMENT_CTX_KEY] ?? null;
-}
-
-export function getBfcacheIdMapContext(): React.Context<Readonly<
-  Record<string, string>
-> | null> | null {
-  if (typeof React.createContext !== "function") return null;
-
-  const globalState = globalThis as _LayoutSegmentGlobal;
-  if (!globalState[_BFCACHE_ID_MAP_CTX_KEY]) {
-    globalState[_BFCACHE_ID_MAP_CTX_KEY] = React.createContext<Readonly<
-      Record<string, string>
-    > | null>(null);
-  }
-
-  return globalState[_BFCACHE_ID_MAP_CTX_KEY] ?? null;
-}
-
-export function getBfcacheSegmentIdContext(): React.Context<string | null> | null {
-  if (typeof React.createContext !== "function") return null;
-
-  const globalState = globalThis as _LayoutSegmentGlobal;
-  if (!globalState[_BFCACHE_SEGMENT_ID_CTX_KEY]) {
-    globalState[_BFCACHE_SEGMENT_ID_CTX_KEY] = React.createContext<string | null>(null);
-  }
-
-  return globalState[_BFCACHE_SEGMENT_ID_CTX_KEY] ?? null;
-}
+// The contexts and request-state bridge live in navigation-context-state.ts so
+// the browser and server facades share one lightweight implementation.
 
 /**
  * Read the child segments for a parallel route below the current layout.
@@ -195,12 +144,6 @@ function useChildSegments(parallelRoutesKey: string = "children"): string[] {
 // Server-side request context (set by the RSC entry before rendering)
 // ---------------------------------------------------------------------------
 
-export type NavigationContext = {
-  pathname: string;
-  searchParams: URLSearchParams;
-  params: Record<string, string | string[]>;
-};
-
 const _READONLY_SEARCH_PARAMS = Symbol("vinext.navigation.readonlySearchParams");
 const _READONLY_SEARCH_PARAMS_SOURCE = Symbol("vinext.navigation.readonlySearchParamsSource");
 const _READONLY_SEARCH_PARAMS_SOURCE_KEY = Symbol(
@@ -212,131 +155,6 @@ type NavigationContextWithReadonlyCache = NavigationContext & {
   [_READONLY_SEARCH_PARAMS_SOURCE]?: URLSearchParams;
   [_READONLY_SEARCH_PARAMS_SOURCE_KEY]?: string;
 };
-
-// ---------------------------------------------------------------------------
-// Server-side navigation state lives in a separate server-only module
-// (navigation-state.ts) that uses AsyncLocalStorage for request isolation.
-// This module is bundled for the browser, so it can't import node:async_hooks.
-//
-// On the server: state functions are set by navigation-state.ts at import time.
-// On the client: _serverContext falls back to null (hooks use window instead).
-//
-// Global accessor pattern (issue #688):
-// Vite's multi-environment dev mode can create separate module instances of
-// this file for the SSR entry vs "use client" components. When that happens,
-// _registerStateAccessors only updates the SSR entry's instance, leaving the
-// "use client" instance with the default (null) fallbacks.
-//
-// To fix this, navigation-state.ts also stores the accessors on globalThis
-// via Symbol.for, and the defaults here check for that global before falling
-// back to module-level state. This ensures all module instances can reach the
-// ALS-backed state regardless of which instance was registered.
-// ---------------------------------------------------------------------------
-
-type _StateAccessors = {
-  getServerContext: () => NavigationContext | null;
-  setServerContext: (ctx: NavigationContext | null) => void;
-  getInsertedHTMLCallbacks: () => Array<() => unknown>;
-  clearInsertedHTMLCallbacks: () => void;
-};
-
-export const GLOBAL_ACCESSORS_KEY = Symbol.for("vinext.navigation.globalAccessors");
-const _GLOBAL_ACCESSORS_KEY = GLOBAL_ACCESSORS_KEY;
-type _GlobalWithAccessors = typeof globalThis & { [_GLOBAL_ACCESSORS_KEY]?: _StateAccessors };
-const _FALLBACK_STATE_KEY = Symbol.for("vinext.navigation.fallback");
-type _FallbackState = {
-  serverContext: NavigationContext | null;
-  serverInsertedHTMLCallbacks: Array<() => unknown>;
-};
-type _GlobalWithFallbackState = typeof globalThis & {
-  [_FALLBACK_STATE_KEY]?: _FallbackState;
-};
-
-// Browser hydration has the same module-split shape as SSR in Vite dev:
-// the browser entry seeds the snapshot before hydrateRoot(), but client
-// components can import a different module instance of this shim.
-const GLOBAL_HYDRATION_CONTEXT_KEY = Symbol.for("vinext.navigation.clientHydrationContext");
-const _GLOBAL_HYDRATION_CONTEXT_KEY = GLOBAL_HYDRATION_CONTEXT_KEY;
-type _GlobalWithHydrationContext = typeof globalThis & {
-  [_GLOBAL_HYDRATION_CONTEXT_KEY]?: NavigationContext | null;
-};
-
-function _getGlobalAccessors(): _StateAccessors | undefined {
-  return (globalThis as _GlobalWithAccessors)[_GLOBAL_ACCESSORS_KEY];
-}
-
-function _getFallbackState(): _FallbackState {
-  const globalState = globalThis as _GlobalWithFallbackState;
-  return (globalState[_FALLBACK_STATE_KEY] ??= {
-    serverContext: null,
-    serverInsertedHTMLCallbacks: [],
-  });
-}
-
-function _getClientHydrationContext(): NavigationContext | null | undefined {
-  const globalState = globalThis as _GlobalWithHydrationContext;
-  if (Object.prototype.hasOwnProperty.call(globalState, _GLOBAL_HYDRATION_CONTEXT_KEY)) {
-    return globalState[_GLOBAL_HYDRATION_CONTEXT_KEY] ?? null;
-  }
-  return undefined;
-}
-
-function _setClientHydrationContext(ctx: NavigationContext | null): void {
-  (globalThis as _GlobalWithHydrationContext)[_GLOBAL_HYDRATION_CONTEXT_KEY] = ctx;
-}
-
-function clearClientHydrationContext(): void {
-  if (typeof window !== "undefined") {
-    _setClientHydrationContext(null);
-  }
-}
-
-// These are overridden by navigation-state.ts on the server to use ALS.
-// The defaults check globalThis for cross-module-instance access (issue #688).
-let _getServerContext = (): NavigationContext | null => {
-  if (typeof window !== "undefined") {
-    const hydrationContext = _getClientHydrationContext();
-    return hydrationContext !== undefined ? hydrationContext : _getFallbackState().serverContext;
-  }
-  const g = _getGlobalAccessors();
-  return g ? g.getServerContext() : _getFallbackState().serverContext;
-};
-let _setServerContext = (ctx: NavigationContext | null): void => {
-  if (typeof window !== "undefined") {
-    _getFallbackState().serverContext = ctx;
-    _setClientHydrationContext(ctx);
-    return;
-  }
-  const g = _getGlobalAccessors();
-  if (g) {
-    g.setServerContext(ctx);
-  } else {
-    _getFallbackState().serverContext = ctx;
-  }
-};
-let _getInsertedHTMLCallbacks = (): Array<() => unknown> => {
-  const g = _getGlobalAccessors();
-  return g ? g.getInsertedHTMLCallbacks() : _getFallbackState().serverInsertedHTMLCallbacks;
-};
-let _clearInsertedHTMLCallbacks = (): void => {
-  const g = _getGlobalAccessors();
-  if (g) {
-    g.clearInsertedHTMLCallbacks();
-  } else {
-    _getFallbackState().serverInsertedHTMLCallbacks = [];
-  }
-};
-
-/**
- * Register ALS-backed state accessors. Called by navigation-state.ts on import.
- * @internal
- */
-export function _registerStateAccessors(accessors: _StateAccessors): void {
-  _getServerContext = accessors.getServerContext;
-  _setServerContext = accessors.setServerContext;
-  _getInsertedHTMLCallbacks = accessors.getInsertedHTMLCallbacks;
-  _clearInsertedHTMLCallbacks = accessors.clearInsertedHTMLCallbacks;
-}
 
 // ---------------------------------------------------------------------------
 // Pages Router compat source.
@@ -362,23 +180,6 @@ const PAGES_NAVIGATION_NOTIFY_KEY = Symbol.for("vinext.navigation.pagesNavigatio
 type _GlobalWithPagesNotify = typeof globalThis & {
   [PAGES_NAVIGATION_NOTIFY_KEY]?: () => void;
 };
-
-/**
- * Get the navigation context for the current SSR/RSC render.
- * Reads from AsyncLocalStorage when available (concurrent-safe),
- * otherwise falls back to module-level state.
- */
-export function getNavigationContext(): NavigationContext | null {
-  return _getServerContext();
-}
-
-/**
- * Set the navigation context for the current SSR/RSC render.
- * Called by the framework entry before rendering each request.
- */
-export function setNavigationContext(ctx: NavigationContext | null): void {
-  _setServerContext(ctx);
-}
 
 // ---------------------------------------------------------------------------
 // Client-side state
@@ -1211,7 +1012,7 @@ let _cachedEmptyClientSearchParams: ReadonlyURLSearchParams | null = null;
  * be visible until the next commit.
  */
 function getSearchParamsSnapshot(): ReadonlyURLSearchParams {
-  if (_getServerContext()) return getServerSearchParamsSnapshot();
+  if (getNavigationContext()) return getServerSearchParamsSnapshot();
 
   const pagesCtx = _getPagesNavigationContext();
   if (pagesCtx) {
@@ -1249,7 +1050,7 @@ function syncCommittedUrlStateFromLocation(): boolean {
 }
 
 function getServerSearchParamsSnapshot(): ReadonlyURLSearchParams {
-  const ctx = _getServerContext() as NavigationContextWithReadonlyCache | null;
+  const ctx = getNavigationContext() as NavigationContextWithReadonlyCache | null;
 
   if (!ctx) {
     // No App Router server context - try Pages Router compat shim.
@@ -1451,7 +1252,7 @@ export function clearPendingPathname(navId: number): void {
 
 function getClientParamsSnapshot(): Record<string, string | string[]> | null {
   const state = getClientNavigationState();
-  const ctx = _getServerContext();
+  const ctx = getNavigationContext();
   if (ctx) return ctx.params;
 
   const pagesCtx = _getPagesNavigationContext();
@@ -1462,7 +1263,7 @@ function getClientParamsSnapshot(): Record<string, string | string[]> | null {
 }
 
 function getServerParamsSnapshot(): Record<string, string | string[]> | null {
-  const ctx = _getServerContext();
+  const ctx = getNavigationContext();
   if (ctx) return ctx.params;
   // No App Router navigation context — fall back to Pages Router state.
   // See `adaptForPathParams` in Next.js's pages-router adapter:
@@ -1492,7 +1293,7 @@ export function usePathname(): string | null {
     markPprFallbackShellDynamicBoundary();
     // During SSR of "use client" components, the navigation context may not be set.
     // Return a safe fallback — the client will hydrate with the real value.
-    const ctx = _getServerContext();
+    const ctx = getNavigationContext();
     if (ctx) return ctx.pathname;
     // Pages Router compat shim: derive pathname from the Pages Router state.
     const pagesCtx = _getPagesNavigationContext();
@@ -1501,7 +1302,7 @@ export function usePathname(): string | null {
   const renderSnapshot = useClientNavigationRenderSnapshot();
   // Client-side: use the hook system for reactivity
   const pathname = React.useSyncExternalStore(subscribeToNavigation, getPathnameSnapshot, () => {
-    const ctx = _getServerContext();
+    const ctx = getNavigationContext();
     if (ctx) return ctx.pathname;
     const pagesCtx = _getPagesNavigationContext();
     return pagesCtx ? pagesCtx.pathname : "/";
@@ -2287,459 +2088,7 @@ export function useServerInsertedHTML(callback: () => unknown): void {
     // Client-side: no-op (styles are already in the DOM)
     return;
   }
-  _getInsertedHTMLCallbacks().push(callback);
-}
-
-/**
- * Flush all collected useServerInsertedHTML callbacks.
- * Returns an array of results (React elements or strings).
- * Clears the callback list so the next render starts fresh.
- *
- * Called by the SSR entry after renderToReadableStream completes.
- */
-export function flushServerInsertedHTML(): unknown[] {
-  const callbacks = _getInsertedHTMLCallbacks();
-  const results: unknown[] = [];
-  for (const cb of callbacks) {
-    try {
-      const result = cb();
-      if (result != null) results.push(result);
-    } catch {
-      // Ignore errors from individual callbacks
-    }
-  }
-  callbacks.length = 0;
-  return results;
-}
-
-/**
- * Render collected useServerInsertedHTML callbacks without unregistering them.
- *
- * Streaming SSR needs to invoke the same style-registry callbacks after each
- * Fizz flush. Libraries such as styled-components and Emotion clear their own
- * per-flush buffers inside the callback; the registration itself must survive
- * until the request stream is closed.
- */
-export function renderServerInsertedHTML(): unknown[] {
-  const callbacks = _getInsertedHTMLCallbacks();
-  const results: unknown[] = [];
-  for (const cb of callbacks) {
-    try {
-      const result = cb();
-      if (result != null) results.push(result);
-    } catch {
-      // Ignore errors from individual callbacks
-    }
-  }
-  return results;
-}
-
-/**
- * Clear all collected useServerInsertedHTML callbacks without flushing.
- * Used for cleanup between requests.
- */
-export function clearServerInsertedHTML(): void {
-  _clearInsertedHTMLCallbacks();
-}
-
-// ---------------------------------------------------------------------------
-// Non-hook utilities (can be called from Server Components)
-// ---------------------------------------------------------------------------
-
-/**
- * HTTP Access Fallback error code — shared prefix for notFound/forbidden/unauthorized.
- * Matches Next.js 16's unified error handling approach.
- */
-export const HTTP_ERROR_FALLBACK_ERROR_CODE = "NEXT_HTTP_ERROR_FALLBACK";
-
-/**
- * Check if an error is an HTTP Access Fallback error (notFound, forbidden, unauthorized).
- */
-export function isHTTPAccessFallbackError(error: unknown): boolean {
-  if (error && typeof error === "object" && "digest" in error) {
-    const digest = String((error as { digest: unknown }).digest);
-    return (
-      digest === "NEXT_NOT_FOUND" || // legacy compat
-      digest.startsWith(`${HTTP_ERROR_FALLBACK_ERROR_CODE};`)
-    );
-  }
-  return false;
-}
-
-/**
- * Extract the HTTP status code from an HTTP Access Fallback error.
- * Returns 404 for legacy NEXT_NOT_FOUND errors.
- */
-export function getAccessFallbackHTTPStatus(error: unknown): number {
-  if (error && typeof error === "object" && "digest" in error) {
-    const digest = String((error as { digest: unknown }).digest);
-    if (digest === "NEXT_NOT_FOUND") return 404;
-    if (digest.startsWith(`${HTTP_ERROR_FALLBACK_ERROR_CODE};`)) {
-      return parseInt(digest.split(";")[1], 10);
-    }
-  }
-  return 404;
-}
-
-/**
- * Enum matching Next.js RedirectType for type-safe redirect calls.
- */
-export enum RedirectType {
-  push = "push",
-  replace = "replace",
-}
-
-/**
- * Internal error class used by redirect/notFound/forbidden/unauthorized.
- * The `digest` field is the serialised control-flow signal read by the
- * framework's error boundary and server-side request handlers.
- */
-class VinextNavigationError extends Error {
-  readonly digest: string;
-  constructor(message: string, digest: string) {
-    super(message);
-    this.digest = digest;
-  }
-}
-
-/**
- * Throw a redirect. Caught by the framework to send a redirect response.
- *
- * When `type` is omitted, the digest carries an empty sentinel so the
- * catch site can resolve the default based on context:
- * - Server Action context → "push"  (Back button works after form submission)
- * - SSR render context    → "replace"
- *
- * This matches Next.js behavior where `redirect()` checks
- * `actionAsyncStorage.getStore()?.isAction` at call time.
- *
- * @see https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/redirect.ts
- */
-export function redirect(url: string, type?: "replace" | "push" | RedirectType): never {
-  throw new VinextNavigationError(
-    `NEXT_REDIRECT:${url}`,
-    `NEXT_REDIRECT;${type ?? ""};${encodeURIComponent(url)}`,
-  );
-}
-
-/**
- * Trigger a permanent redirect (308).
- *
- * Accepts an optional `type` parameter matching Next.js's signature.
- * Defaults to "replace" (not context-dependent like `redirect()`).
- *
- * @see https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/redirect.ts
- */
-export function permanentRedirect(
-  url: string,
-  type: "replace" | "push" | RedirectType = "replace",
-): never {
-  throw new VinextNavigationError(
-    `NEXT_REDIRECT:${url}`,
-    `NEXT_REDIRECT;${type};${encodeURIComponent(url)};308`,
-  );
-}
-
-/**
- * Trigger a not-found response (404). Caught by the framework.
- */
-export function notFound(): never {
-  throw new VinextNavigationError("NEXT_NOT_FOUND", `${HTTP_ERROR_FALLBACK_ERROR_CODE};404`);
-}
-
-/**
- * Trigger a forbidden response (403). Caught by the framework.
- * In Next.js, this is gated behind experimental.authInterrupts — we
- * support it unconditionally for maximum compatibility.
- */
-export function forbidden(): never {
-  throw new VinextNavigationError("NEXT_FORBIDDEN", `${HTTP_ERROR_FALLBACK_ERROR_CODE};403`);
-}
-
-/**
- * Trigger an unauthorized response (401). Caught by the framework.
- * In Next.js, this is gated behind experimental.authInterrupts — we
- * support it unconditionally for maximum compatibility.
- */
-export function unauthorized(): never {
-  throw new VinextNavigationError("NEXT_UNAUTHORIZED", `${HTTP_ERROR_FALLBACK_ERROR_CODE};401`);
-}
-
-// ---------------------------------------------------------------------------
-// Internal-error predicates and rethrow
-//
-// `unstable_rethrow` is part of Next.js's public API. User code in try/catch
-// wrappers calls it to let Next.js's control-flow signals (redirect, notFound,
-// forbidden, unauthorized, dynamic-server-usage, bailout-to-CSR, …)
-// propagate up to the framework instead of being swallowed. The canonical
-// use case is a `fetch()` retry helper that needs to bail out the moment
-// fetch throws a framework signal — see Next.js's
-// test/e2e/app-dir/app-static/lib/fetch-retry.js.
-//
-// Ported from Next.js:
-//   - packages/next/src/client/components/unstable-rethrow.ts (dispatcher)
-//   - packages/next/src/client/components/unstable-rethrow.browser.ts
-//   - packages/next/src/client/components/unstable-rethrow.server.ts
-//   - packages/next/src/client/components/is-next-router-error.ts
-//   - packages/next/src/client/components/redirect-error.ts
-//   - packages/next/src/shared/lib/lazy-dynamic/bailout-to-csr.ts
-//   - packages/next/src/client/components/hooks-server-context.ts
-//
-// Coverage of Next.js's 7 server-side categories (server build):
-//   ✓ isNextRouterError (#1) — redirect + HTTP access fallback
-//   ✓ isBailoutToCSRError (#2) — digest === "BAILOUT_TO_CLIENT_SIDE_RENDERING"
-//   ✓ isDynamicServerError (#3) — digest === "DYNAMIC_SERVER_USAGE"
-//   ✗ isDynamicPostpone (#4) — PPR-internal message check; vinext has no PPR
-//   ✗ isPostpone (#5) — React.unstable_postpone signal; vinext has no PPR
-//   ✗ isHangingPromiseRejectionError (#6) — prerender abort signal
-//   ✗ isPrerenderInterruptedError (#7) — prerender controller interrupt
-//
-// The four uncovered categories are server-only Next.js internals tied to
-// prerender-machinery vinext does not implement; user code cannot construct
-// them in normal use. They will be added if/when vinext grows PPR support.
-// ---------------------------------------------------------------------------
-
-type _RedirectErrorShape = Error & { digest: string };
-
-/**
- * Check whether an error was produced by `redirect()` or `permanentRedirect()`.
- *
- * **Note on vinext public surface:** Next.js does NOT expose `isRedirectError`
- * from `next/navigation` — it's an internal predicate. vinext exposes it for
- * symmetry with the already-public `isHTTPAccessFallbackError` and because
- * `unstable_rethrow` consumers benefit from being able to narrow types.
- * Treat it as a vinext-only extension.
- *
- * **Divergence from Next.js:** Next.js's internal `isRedirectError` performs
- * full 4-segment validation — it splits the digest on `;`, checks `type` ∈
- * {push, replace}, requires a non-empty destination, and validates the
- * status code (303, 307, 308). See:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/redirect-error.ts
- *
- * vinext instead uses a simple prefix check (`startsWith("NEXT_REDIRECT;")`).
- * Reasons:
- *   1. vinext emits two digest shapes — 3-part for `redirect()`
- *      (`NEXT_REDIRECT;{type};{encoded-url}`) and 4-part for
- *      `permanentRedirect()` (`NEXT_REDIRECT;{type};{encoded-url};308`).
- *      Strict validation would have to special-case both, and Next.js's
- *      validator (tuned to its 5-part canary digests) rejects them.
- *   2. The `type` field is sometimes empty in vinext's redirect digests
- *      (context-dependent resolution; see `redirect()` above), which the
- *      strict check disallows.
- *
- * **Consequence:** A malformed digest such as `"NEXT_REDIRECT;garbage"`
- * returns `true` here, whereas Next.js would return `false`. In practice,
- * the only callers of this predicate are vinext-internal code paths
- * (`unstable_rethrow`, `unstable_catchError`, the redirect error boundary)
- * that see digests vinext itself emits — so the divergence does not surface
- * in normal use. Maintainers extending the prefix logic should keep this
- * predicate in lockstep with the corresponding `decode*` helpers in
- * `shims/error-boundary.tsx`.
- */
-export function isRedirectError(error: unknown): error is _RedirectErrorShape {
-  if (!error || typeof error !== "object") return false;
-  if (!("digest" in error)) return false;
-  if (typeof error.digest !== "string") return false;
-  return error.digest.startsWith("NEXT_REDIRECT;");
-}
-
-/**
- * Parse a redirect error digest into its URL and type components.
- *
- * Supports two formats:
- *   - vinext's 3-part: `NEXT_REDIRECT;{type};{encoded-url}`
- *   - Next.js's 5-part: `NEXT_REDIRECT;{type};{url};{status};{isClient}`
- *
- * The URL segment is always percent-encoded on the write side
- * (encodeURIComponent is used), so re-joining with ";" for the 5-part
- * format is defensive — it correctly handles any unencoded ";" that
- * might appear in an externally-sourced digest.
- *
- * Returns null for malformed digests that have an empty URL segment, or
- * when the URL contains invalid percent-encoding.
- */
-export function decodeRedirectError(
-  digest: string,
-): { url: string; type: "push" | "replace" } | null {
-  if (!digest.startsWith("NEXT_REDIRECT;")) return null;
-
-  const parts = digest.split(";");
-  const encodedTarget = parts.length >= 5 ? parts.slice(2, -2).join(";") : parts[2];
-  if (!encodedTarget) return null;
-
-  let url: string;
-  try {
-    url = decodeURIComponent(encodedTarget);
-  } catch {
-    return null;
-  }
-
-  const type: "push" | "replace" = parts[1] === "push" ? "push" : "replace";
-  return { url, type };
-}
-
-/**
- * Returns true if the error is a Next.js navigation signal — either a redirect
- * or an HTTP access fallback (notFound / forbidden / unauthorized).
- *
- * **Note on vinext public surface:** Like `isRedirectError`, Next.js does NOT
- * expose this from `next/navigation`. vinext exposes it for symmetry — treat
- * it as a vinext-only extension.
- *
- * Ported from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/is-next-router-error.ts
- */
-export function isNextRouterError(error: unknown): boolean {
-  return isRedirectError(error) || isHTTPAccessFallbackError(error);
-}
-
-// ---------------------------------------------------------------------------
-// BailoutToCSRError — `next/dynamic` with `ssr: false` throws this during
-// server render to signal that the dynamic component must be rendered on
-// the client. Lives in shared (non-server) code so it can flow through both
-// the SSR pipeline and userland; third-party libraries that emulate
-// `next/dynamic` also construct it.
-//
-// Ported from Next.js:
-//   https://github.com/vercel/next.js/blob/canary/packages/next/src/shared/lib/lazy-dynamic/bailout-to-csr.ts
-// ---------------------------------------------------------------------------
-
-const _BAILOUT_TO_CSR_DIGEST = "BAILOUT_TO_CLIENT_SIDE_RENDERING";
-
-/**
- * Error thrown to bail out of server rendering and fall back to client-side
- * rendering. Used by `next/dynamic` with `ssr: false`.
- *
- * vinext does not yet emit this error itself — it's exposed so user code and
- * third-party libraries that mimic `next/dynamic`'s bailout semantics can
- * construct an error with the canonical digest that `unstable_rethrow`
- * recognises.
- *
- * Ported 1:1 from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/shared/lib/lazy-dynamic/bailout-to-csr.ts
- */
-export class BailoutToCSRError extends Error {
-  public readonly digest: typeof _BAILOUT_TO_CSR_DIGEST = _BAILOUT_TO_CSR_DIGEST;
-  public readonly reason: string;
-
-  constructor(reason: string) {
-    super(`Bail out to client-side rendering: ${reason}`);
-    this.reason = reason;
-  }
-}
-
-/**
- * Returns true if the error is a `BailoutToCSRError`. Matches Next.js's
- * digest-based predicate, so any error from a foreign module instance of
- * the class (or constructed manually with the canonical digest) is also
- * detected.
- *
- * **Note on vinext public surface:** Next.js does NOT expose this from
- * `next/navigation`. vinext exposes it for symmetry with `isRedirectError`
- * — treat it as a vinext-only extension. The matching producer
- * (`BailoutToCSRError`) is the public detection contract; Next.js exposes
- * neither.
- *
- * Ported from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/shared/lib/lazy-dynamic/bailout-to-csr.ts
- */
-export function isBailoutToCSRError(error: unknown): error is BailoutToCSRError {
-  if (!error || typeof error !== "object" || !("digest" in error)) {
-    return false;
-  }
-  return (error as { digest: unknown }).digest === _BAILOUT_TO_CSR_DIGEST;
-}
-
-// ---------------------------------------------------------------------------
-// DynamicServerError — thrown by Next.js's internal `cookies()`/`headers()`
-// shims when called inside a static render context that cannot resolve
-// request-scoped data. vinext's own `next/headers` shim has its own throw
-// semantics, so vinext never constructs this error itself, but third-party
-// code or accidentally-bundled Next.js internals can.
-//
-// Ported from Next.js:
-//   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/hooks-server-context.ts
-// ---------------------------------------------------------------------------
-
-const _DYNAMIC_SERVER_USAGE_DIGEST = "DYNAMIC_SERVER_USAGE";
-
-/**
- * Error thrown when dynamic server APIs (`cookies()`, `headers()`, etc.) are
- * used inside a static/prerender context. Carries the `DYNAMIC_SERVER_USAGE`
- * digest so `unstable_rethrow` can recognise and propagate it.
- *
- * vinext does not construct this error itself — exposed for the same
- * "stable detection contract" reason as `BailoutToCSRError` above.
- *
- * Ported 1:1 from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/hooks-server-context.ts
- */
-export class DynamicServerError extends Error {
-  public readonly digest: typeof _DYNAMIC_SERVER_USAGE_DIGEST = _DYNAMIC_SERVER_USAGE_DIGEST;
-  public readonly description: string;
-
-  constructor(description: string) {
-    super(`Dynamic server usage: ${description}`);
-    this.description = description;
-  }
-}
-
-/**
- * Returns true if the error is a `DynamicServerError` (or any error with the
- * canonical `DYNAMIC_SERVER_USAGE` digest).
- *
- * **Note on vinext public surface:** Next.js does NOT expose this from
- * `next/navigation`. vinext exposes it for symmetry — treat it as a
- * vinext-only extension.
- *
- * Ported from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/hooks-server-context.ts
- */
-export function isDynamicServerError(error: unknown): error is DynamicServerError {
-  if (!error || typeof error !== "object" || !("digest" in error)) {
-    return false;
-  }
-  // `===` against a string literal already requires the operand to be a
-  // string, so no separate `typeof digest === "string"` check is needed.
-  // Matches `isBailoutToCSRError` above for stylistic consistency.
-  return (error as { digest: unknown }).digest === _DYNAMIC_SERVER_USAGE_DIGEST;
-}
-
-/**
- * Rethrow internal Next.js errors so they're handled by the framework.
- *
- * When wrapping an API that uses errors for control flow (redirect, notFound,
- * cookies in static render, `next/dynamic` SSR bailout, etc.), call this
- * inside `catch` blocks before doing your own error handling. If the error
- * is a Next.js internal error, it's rethrown; otherwise this is a no-op
- * (apart from recursing through `error.cause`).
- *
- * Recognises (matches Next.js's browser build + the subset of the server
- * build that vinext can realistically encounter):
- *   - `isNextRouterError`: redirect / notFound / forbidden / unauthorized
- *   - `isBailoutToCSRError`: `next/dynamic` `ssr: false` bailout
- *   - `isDynamicServerError`: dynamic API used in static render
- *
- * vinext does not yet recognise four additional server-only Next.js
- * categories — `isDynamicPostpone`, `isPostpone`,
- * `isHangingPromiseRejectionError`, `isPrerenderInterruptedError` — because
- * they signal PPR / prerender-controller events that vinext's render
- * pipeline does not generate. User code cannot construct these in normal
- * use; they will be added if/when vinext grows PPR support.
- *
- * Ported from Next.js:
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/unstable-rethrow.ts
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/unstable-rethrow.server.ts
- *   https://github.com/vercel/next.js/blob/canary/packages/next/src/client/components/unstable-rethrow.browser.ts
- */
-export function unstable_rethrow(error: unknown): void {
-  if (isNextRouterError(error) || isBailoutToCSRError(error) || isDynamicServerError(error)) {
-    throw error;
-  }
-
-  if (error instanceof Error && "cause" in error) {
-    unstable_rethrow((error as Error & { cause: unknown }).cause);
-  }
+  registerServerInsertedHTMLCallback(callback);
 }
 
 // ---------------------------------------------------------------------------
